requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.0
pydantic>=2.10.6
sqlalchemy>=2.0.39
psycopg>=3.2.6
redis>=5.2.1
bcrypt>=4.3.0
python-dotenv>=1.1.0
python-jose>=3.4.0

# testing
pytest>=8.3.5
pytest-asyncio>=0.26.0
fakeredis>=2.28.0
httpx>=0.28.1
locust>=2.33.2
pytest-cov>=6.0.0
